# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Esse é a API do BQ\n",
    "version": "1.0.0",
    "title": "Burger Queen API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "monicay@laboratoria.la"
    }
  },
  "host": "lab-api-bq.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/orders/{orderId}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get order by Id",
        "operationId": "getOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Auth token",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Orders"
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "Update order",
        "description": "Update order status",
        "operationId": "updateOrder",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Auth token",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "ID that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "the new order status",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Orders"
            }
          },
          "400": {
            "description": "Missing required data or no changes applied"
          },
          "401": {
            "description": "User not authorized"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "delete": {
        "tags": [
          "orders"
        ],
        "summary": "Delete order",
        "description": "Delete order",
        "operationId": "deleteOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Auth token",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "The order that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Orders"
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Orders": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "client_name": {
          "type": "string"
        },
        "user_id": {
          "type": "integer",
          "format": "int64"
        },
        "table": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "processedAt": {
          "type": "date",
          "description": "Date of the first update process"
        },
        "createdAt": {
          "type": "date",
          "description": "Date of the creation process"
        },
        "updatedAt": {
          "type": "date",
          "description": "Date of the last update process"
        },
        "Products": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "name": {
                "type": "string"
              },
              "flavor": {
                "type": "string"
              },
              "complement": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  }
}